Initial version, with a reasonable watch face and a companion app
This initial version contains a reasonable watch face with support
for showing time reasonably, an optional background, support for
ambient and low-bit modes and timezones. There is a basic companion
app with a simple activity and simple message passing between the
companion app and the watch face. As such, we now want to add
the intelligence.

This is directly descended from the WatchFace sample. As such, it
bears the mark with some remaining code that is now either useless or
not well adapted to what we want to do. Because of this some cleanup
and refactoring and renaming is in order, though a good deal of it has
been done already.