Create the new independent client.
This is an initial, untested implementation. The basic structure is
that the Client class implements a stateless interface to a stateful
asynchronous client using a Handler thread, an action queue and an
event model.

This change implements the first action, PutDataAction, alongside
with the Client.

Next steps are : first, probably convert to a quasi-procedural
style in the handleMessage method and have a proceed() method
that figures out the right message. That's going to be easier
to read. Then debug.