Allow use of annotations in the Common module
This is an ugly solution to this problem - it won't work outside
of my local system, and will fail on another machine. Still, for
the time being I am fed up and this works, and it is better than
not being able to use annotations in Common. Let's refine later.