Route the sound into the text fields.
Also remove the recognition listener, merge it in
the sound source, as it had no value on its own.

There are still a few bugs. Importantly, the router
needs to inform the source of focus/cursor changes so
that the source knows not to continue delivering
partial or final results of the same utterance.

...and ideally we would not be listening while not
in a field, unless we want to stream commands in
this case.

Also, still no idea why sound levels work every
fourth time, but I fear this may be a bug on their
side and we can't do anything. Oh well, it works
even without the levels, it does not matter too
much.