Add a partial recursive backup action.
The recursive backup action for now only resolves and/or creates
the directories as needed. Subsequent changes will implement the
backup itself.

This takes good advantage of the recent dependent action facility.
The code for the backup action is quite satisfactory (though not
perfect) and the code for resolving directories is not as good as
I hoped but it works the right way.