Change Action to support dependencies and add ResultAction.
Unfortunately that means two things :
- The client needs to be passed to the constructor of every Action.
  It's a bit sad, but it was already the case for some anyway, and
  it's more explicit that way. It lets clients call .enqueue() on
  the action directly, which also has advantages when passing the
  action to someone else.
- Actions have to call finish() when they're done. Unless they
  have no dependency, in which case it doesn't matter, but it's not
  very comfortable because typically an Action is not supposed to
  know if it has any dependency. In the practice the dependency
  is passed as an arg constructor and it's final, so if passed null
  it's safe, and anyone who needs a dependency on that action has
  to modify that code, so is bound to notice. But still.

Of course the point of doing this is that advantages outweigh
drawbacks. Now any Action can re-enqueue itself in a standard way,
instead of each Action that wants to do this having to embark the
client as a member. Dependencies should work out of the box (but
is of course untested as of this change). And the addition of
ResultAction on top of this as an implementor of Future means a
standard, clean and easy way to implement dependent classes that
need a result out of their dependencies. It's pretty neat.