Allow moving todos by dragging them.
This does not work as expected for TODOs with children yet,
it moves the parent and orphans the children. It kind of sucks.

However, a good refactoring is in order now that I have a better
idea of where to go (in particular the list of shown indices).
Fixing this bug can probably wait until the code is clearer and
easier to work with in these respects.