Have a Todo embark a TodoUI.
This simplifies things vastly, and honestly I'm a lot more
at ease with this guaranteeing copying Todos copy their
UI params and guaranteeing there won't be a case where
forgetting to put a TodoUI into the hash will cause a
crash later. Also, there is no need to clean up the hash,
which wasn't even done. All the important parts are still
immutable, the mutable parts are encapsulated, and there
is still a truly immutable object underneath that can be
gotten to if needed.

It's Better this way.