Update the app for Rio's new work and home.
This entailed writing a new parser. Ekitan is the site I chose.
It's really nice because on top of having stations/lines from
many companies (which will likely be very helpful on every
future update which changes stations), it also has an easy to
parse format, and each train even has a link to when it reaches
where which may be useful in the future.

Also this centers the departures and headsign in the watch.
That was a much better choice with long departures on a small
watch screen.

This temporarily disables the backup action call.