Implement the recursive backup.
With this, the DB is sync'd to GDrive with every change after 3 seconds,
and files are backed up regularly in case anything ever goes wrong.

This is not well tested yet, but as of this writing there is no known
bug. In particular, some things will only be known in two weeks :)