Branch for Rio mode.
Put a cat as the background picture, remove Drive permission,
make a more evocative app icon.

This branch may be short-lived ; after all it won't be tracking
trunk or contributing to it, or at least that's not the plan.
It's here for future reference of what needs to be done, so that
it can either be done again, or summarily built upon, or
generalized so it won't have to diverge, whatever.