Add the image editor. Not really an editor but.
Choose an image, display it, scroll it, zoom it and pan it, and
choose a place to crop. Most of the work is done courtesy of
Ortiz Michael's TouchImageView, with a very small change of mine
to avoid complex computation.

Some small cleanup too, like autoconvert of ugly inline classes to
lambda goodness.