Fix the tests for moving.
This finally has functional unit tests for moving Todos.
The test will read the fixture DB for status and collect
all expandable Todos. It will generate all possible states
of the UI (64 at this point because the fixture contains
6 different expandable Todos). For each of these states,
it will simulate moving every Todo to every possible
destination.

This change fixes many bugs in the test and refactors it
extensively for readability. It also fixes a small bug in
the main program (exactly the kind of bug that this test
was written to find).