Migrate GeofenceAction (not running yet). Also InformUserAction.
Migrate GeofenceAction from the old, general Client-I-can-disconnect-
you-and-crash-you-without-warning-whenever-the-fuck-I-want to the
newer one-client-per-API style, with our new Actions.

As the new Action API is yet incomplete and still won't actually run
stuff, this is called but not run. The main crux of the issue is
in Action#enqueue() which right now says TODO : implement me !

While we're at it implement an InformUserAction, which is just a
virtual class returning either a SnackbarAction if there is any
way to show a snackbar right now, or a NotificationAction
otherwise. The current implementation however may not be enough,
because IIUC if the app is backgrounded between when the action
is created and when it's complete, we'll still find an instance
of Snackbarable, and "show" it in a UI that's not on screen.
This should be fixed, dude.