Full rework of the Client API.
Because the GoogleApiClient API is deprecated in favor of
separate *Client for each API, the old Client API is not
useful as its main focus is to make sure the GoogleApiClient
is connected and the new API does not have this utter
nonsense any more.

Still, something is needed for easy writing and reading of
data to Wear. But in many cases it's just helper functions,
or at most a thin veneer to supply an executor to do the
stuff.

This implements a nice way to direct feedback from the app
either to a snackbar or a notification. Notifications work ;
didn't try snackbar yet.

Also, now Fragments can't be anonymous, because Google says
they can't and they check for it. No idea what kind of bug
they hope anyone catches with this, but I can tell you why
it's annoying and not very smart. Anyway, this includes a
workaround for that new behavior, based on just passing the
Class corresponding to the concrete wrapped fragment. That
came together with some improvement in type safety of the
existing WrapUtils utilities, because why not while I'm at
it.

Also minor changes like adding words to the dictionary and
whatnot.

This removes usage for most (all ?) the old Client API as
far as Wearable is concerned, but not the functions or
API themselves. We'll do that soon, and this change is
more than big enough already.

By the way, this time it's tested. It works.