First part of the implementation of TapControl for the watch.
This is incomplete (and won't even build, not least of which
because of intentional markers to show where the code is not
going to work because it's left half-done), but I still want
to commit now to change machines. Also, while "incomplete",
it's not very very far (but of course completely untested).