Tap support for the watch face.
This is rudimentary, but useful. Top quadrant will change the Status,
that is the location from the user's point of view. Left and right
go to the prev / next departures, respectively. Bottom will show / hide
the topic, and reset location and departure time to here and now. Reset
also happens automatically after 5 minutes.