Move from Snackbarable to SnackbarRegistry.
This allows geofences or other possible future stuff to show their
transient info in a snackbar instead of a notification when possible,
while still showing a notification when not possible. It also fixes
the previous regular use case which was broken in the previous impl
(because the Context was not Snackbarable, only the WrappedFragment
which is not a Context).