Implement wear data sync.
The important part here is that all wear data is sync'd correctly
and mirrored to the watch from any device that sets it.
But this change also implements real time sync on the host app
in messages, so that if both phones have it open at the same
time, changes on one are reflected on the other. The last one
wins and I don't give a damn about more than this.
Real time sync on the host app in debug tools is not implemented.
I think I'll do it anyway for the lulz, but it's not like it's
actually useful because the data itself is correctly mirrored
and correctly used on the watch ; for debug data even that
wouldn't matter too much, but it works. The rest is really for
the lulz.