Fix geofences and wear messages.
Revert to an older SDK, because apparently geofences are subject
to background limits under O MR1. Which contradicts the doc.
Right now I see no way out of this, so to have it work for now
revert.
Also move the write to the database to such a time that there
is some network. That lets the code register for a write listener
and send the FCM message (made "urgent" because it needs to reach
even if the remote device's screen is off) when the write to the
DB has been committed, fixing the issue where the message would
arrive and the DB still not be updated.