Add a draft NotifEngine.
It's not functional, I'm starting experimentation is all.

Also add some debug tools to help. My last install of the app was on
another computer so I'd need to log back in after reinstalling the app
with a different cert, but I can't do that because I'm on a plane.