Fugbix. Yes, fugbix.
Fencepost :( When clearing the first todo, its index is 0,
which is >= 0 but not > 0, so the code would try to remove
something at index -1, 'cause I'm stupid like that.
That fug is now bixed.

Also, add comments. 'cause you know I found this code was
not as obvious as I probably thought it was when I wrote it.