Create the split notification and start wiring it down.
This is a draft/exploration at this point. The debug activity
has a button that makes a notification suggesting to split a
(phony) todo with a remote input to write the different items.
The input is taken to be split on comma, which I may change
later because it kind of sucks. For now the data is wired to
JOrg, but not used.