Bugfixes
The intent fired from the launcher is not appropriate to
launch the target of the trampolines : fix this.
Also, in some (unclear) cases the bundle is null (though
it seems to me it's documented not to be ?) and Kotlin
(kind of stupidly ? I see advantages too, but in this case
it's rather annoying) decides to add a null check on an
arg that is never used.