Fix a bug with a todo removal.
The app never does removal (it only marks as done), but it's
possible to do from the web interface and in this case it would
crash the app. Besides maybe the app does want to do removal in
the future.

Also some small code style/comment wording fixes, and adding
some small logging to help diagnose FCM not having the intended
effect (in this case, the key had been mistakenly removed from
the DB ; added some comments to that effect).