Implement splitting a todo directly from the notification
The UX is not great : it requires typing the todos with commas
to separate them.
This sends the string to a receiver which bounces it to a job.
The job creates an instance of a todo list, splits the subitems
on comma, strips them, and creates the subtodos.

As a minor drive-by refactoring, create a new CommonObjects class to
host a common static executor that's created on demand to use from
essentially anywhere.

The elephant in the room after this is that the notification is
never dismissed.