Bugfixes with server updates
Now the server sends a REMOVED update instead of a completionTime
set to 0 update (this is because the query is now limited to todos
with completionTime = 0 for bandwidth reasons). This required
a new path for removal without marking completion – indeed,
without knowing the new todo contents. But that does not matter :
the server says removed, so it gets removed.
Also fix a bug on the converse side : when a new todo is added
with index 0, it used to crash because of a fencepost bug. How
many incarnations of this bug will I have >.>