Update IDE and libs to all latest versions, and upgrade to androidx.
Not going to the betas and RCs, only the latest stable stuff. It's
already annoying and time-consuming enough, even with a relatively
small app like this, to update all dependencies and everything that
has changed ; here I only wanted to upgrade to Crashlytics because
its predecessor has been deprecated, and that was a multiple-hour
stackoverflow fest because of all the stuff Google has broken
because they forgot how developers used their stuff one week ago.