Improve burn-in protection
(though it's not clear it's really useful, I haven't seen it
kick in ever)

Burn-in protection shifts the views around a bit, and the doc says
the watch faces are supposed to avoid drawing anything interesting
in the 10 pixels close to the border when it's on – which is kind
of ridiculous for many faces, but mine can do that by just not
drawing the date when in ambient + burn-in protection.
The doc also says to not shot large expanses of white, so making
the time bold is unambiguously the wrong answer, I don't know why
I did that.

Also smalll readability improvements and make sure the phone
watch face updates when one of the modes is switched or data is
updated, rather than on the next second tick.