Add a new LocalLog facility. Don't load when no net.
The new LocalLog facility is implemented over a Room database.
It is meant to keep a list of text messages that can be later
displayed in UI, even days later. Entries older than 30 days
are removed before they are displayed to avoid blowing up mem.

Also add UI to show and clear it in logs & debug tab.

To do this correctly with coroutines, add lifecycle owners lib,
which required switching to the support fragments and activities.
Surprisingly this was relatively painless, only one thing that
blew up on startup. Maybe some hidden things lurk in the dark,
waiting to trip the app up, bidding their time.

Finally, because I'm sick of the "can't load" error message
when in the train, don't even try to load when there is no net.
Still leave a log it the local log so I can check this is indeed
working as I expect.