Add the "checkpoints" feature.
Double-tapping on the lower quadrant on the watch will
add a "checkpoint", that is, a string with the date and
time of the tap to a list of strings. That list will
then be displayed (and editable) by the phone in the
messages tab, and the phone will put up a notification
with the contents of the list.

The point of this is to let the user put a checkpoint at
some given time so the machine remembers the time instead
of them, together with a notification so they don't forget
they put a reminder for themselves. They can then easily
add details on the phone (for the notification immediately
opens the right activity) if they want to add more context
in case they forget. The notification is non-intrusive,
won't vibrate, and won't show up on the watch (I'm not
sure that's right, but for sure it should not peek
immediately on double tap, that'd be obnoxious).

Improvements may include :
• Add some feedback on the watch that the double tap was
  indeed heared. Right now you don't know until you look
  at the phone, which is worrying.
• Add a delete button in the notification, since right now
  the only way to clear the string is to open the app and
  manually delete the contents of the string.
• Add a text field on the notification to let the user add
  some text without even opening the app.

This version also includes a minor bug fix where the app
would not open again with the last open tab, and updates
for the newest gradle plugin and crashlytics version.