Fix a recently introduced crash
I had figured this out before, but forgot about it so
I restored the code to what is generally advised.
Except this puts in play broken framework mechanics
that the app explicitly refuses to play with, and
crashes the app.
Restoring to the working code, and add comments in the
relevant place to avoid making the same mistake again.