Okay okay it's not great to have an exposed key
Delete it and create a new one that won't go into the repo
Note to future self : if the file is lost, just go create a new key
at the json format and put it at the root of the project (where
was the file deleted by this change). It doesn't look like anything
more is needed ? Not too sure.