Fix reading 京成線
Ekitan changed the format for keisei.ekitan.com. It's good ; they finally
switched to UTF-8 (welcome to this century) and changed a couple
things breaking the old Keisei Ekitan parser. But luckily, now the
regular Ekitan URL also covers 京成線 (which, if I recall, didn't use to
be the case – or if it had been, then I was an idiot for writing a
separate parser), meaning now the regular Ekitan parser can be
used for 京成線. This means there is now only one parser for every
line !

Migrate DataSource and EkitanParser to Kotlin, and have a few
improvements in EkitanParser to support the 京成線-only features
(specifically, the ability to drop some departures because the
train won't stop at 千住大橋). Also genericize the tab number,
which is much nicer.